+++ src/cell.hpp
@@ -26,8 +26,8 @@ public:
 	};
 
 	Cell();
-	Cell(const int &id);
-	Cell(const int &id, Patch *patch);
+	Cell(const long &id);
+	Cell(const long &id, Patch *patch);
 
 	Cell(Cell&& other) = default;
 	Cell& operator=(Cell&& other) = default;
@@ -38,28 +38,28 @@ public:
 	void set_volume(double * const volume);
 	const double & get_volume() const;
 
-	void initialize_interfaces(std::vector<std::vector<int>> &interfaces);
+	void initialize_interfaces(std::vector<std::vector<long>> &interfaces);
 	void initialize_empty_interfaces(const int nInterfaces[]);
-	void set_interface(const int &face, const int &index, const int &interface);
-	void set_interfaces(const int &face, int interface[]);
+	void set_interface(const int &face, const int &index, const long &interface);
+	void set_interfaces(const int &face, long interface[]);
 	void unset_interfaces();
 	int get_interface_count() const;
 	int get_interface_count(const int &face) const;
-	int get_interface(const int &face, const int &index = 0) const;
-	const int * get_interfaces() const;
-	const int * get_interfaces(const int &face) const;
+	long get_interface(const int &face, const int &index = 0) const;
+	const long * get_interfaces() const;
+	const long * get_interfaces(const int &face) const;
 
-	std::vector<int> extract_neighs(int codimension, bool complete = true) const;
+	std::vector<long> extract_neighs(int codimension, bool complete = true) const;
 
-	std::vector<int> extract_face_neighs() const;
-	std::vector<int> extract_face_neighs(const int &face, const std::vector<int> &blackList = std::vector<int>()) const;
+	std::vector<long> extract_face_neighs() const;
+	std::vector<long> extract_face_neighs(const int &face, const std::vector<long> &blackList = std::vector<long>()) const;
 
-	std::vector<int> extract_edge_neighs(bool complete = true) const;
-	std::vector<int> extract_edge_neighs(const int &edge, const std::vector<int> &blackList = std::vector<int>()) const;
+	std::vector<long> extract_edge_neighs(bool complete = true) const;
+	std::vector<long> extract_edge_neighs(const int &edge, const std::vector<long> &blackList = std::vector<long>()) const;
 
-	std::vector<int> extract_vertex_neighs(bool complete = true) const;
-	std::vector<int> extract_vertex_neighs(const int &vertex, const std::vector<int> &blackList = std::vector<int>()) const;
-	std::vector<int> extract_vertex_neighs(const std::vector<int> &vertices, const std::vector<int> &blackList = std::vector<int>()) const;
+	std::vector<long> extract_vertex_neighs(bool complete = true) const;
+	std::vector<long> extract_vertex_neighs(const int &vertex, const std::vector<long> &blackList = std::vector<long>()) const;
+	std::vector<long> extract_vertex_neighs(const std::vector<int> &vertices, const std::vector<long> &blackList = std::vector<long>()) const;
 
 	void set_data(std::unique_ptr<CellData> m_data);
 	CellData * get_data() const;
@@ -71,7 +71,7 @@ private:
 
 	double *m_volume;
 
-	std::unique_ptr<CollapsedArray2D<int> > m_interfaces;
+	std::unique_ptr<CollapsedArray2D<long> > m_interfaces;
 
 	std::unique_ptr<CellData> m_data;
